+++ src/registers.h
@@ -40,6 +40,6 @@ typedef struct Register {
 
 typedef struct bit {unsigned x:1;} bit; /**< bit Data Type */
 
-void printRegister(double inst, int i, pid_t pid, long tid,  int dur, int pl, char *oper);
+void printRegister(long int inst, int i, pid_t pid, long tid,  int dur, int pl, char *oper);
 
 #endif /*REGISTER_H*/
